Trouper, TrouperMap, and SocketTrouper
